Talk architecture polishing.
