+++ giskard/llm/talk/tools/__init__.py
@@ -1,7 +1,7 @@
-from giskard.llm.talk.tools.base import BasePredictTool, BaseTool
+from giskard.llm.talk.tools.base import BaseTool
 from giskard.llm.talk.tools.metric import MetricTool
 from giskard.llm.talk.tools.predict import PredictTool
 from giskard.llm.talk.tools.scan import IssuesScannerTool
 from giskard.llm.talk.tools.shap import SHAPExplanationTool
 
-__all__ = ["BaseTool", "BasePredictTool", "PredictTool", "MetricTool", "IssuesScannerTool", "SHAPExplanationTool"]
+__all__ = ["BaseTool", "PredictTool", "MetricTool", "IssuesScannerTool", "SHAPExplanationTool"]
