+++ giskard/llm/talk/tools/predict.py
@@ -3,17 +3,41 @@ from thefuzz import fuzz
 
 from giskard.datasets.base import Dataset
 from giskard.llm.talk.config import ToolDescription
-from giskard.llm.talk.tools.base import BasePredictTool
+from giskard.llm.talk.tools.base import BaseTool, get_feature_json_type
 
 
-class PredictTool(BasePredictTool):
+class PredictTool(BaseTool):
     default_name: str = "predict"
     default_description: str = ToolDescription.PREDICT.value
 
-    def _filter_from_dataset(self, row_filter: dict) -> pd.DataFrame:
-        threshold = 85
+    @property
+    def specification(self) -> str:
+        feature_json_type = get_feature_json_type(self._dataset)
 
+        return {
+            "type": "function",
+            "function": {
+                "name": self.name,
+                "description": self.description,
+                "parameters": {
+                    "type": "object",
+                    "properties": {
+                        "features_dict": {
+                            "type": "object",
+                            "properties": {
+                                feature: {"type": dtype} for feature, dtype in list(feature_json_type.items())
+                            },
+                        }
+                    },
+                    "required": ["features_dict"],
+                },
+            },
+        }
+
+    def _get_input_from_dataset(self, row_filter: dict) -> pd.DataFrame:
+        threshold = 85
         filtered_df = self._dataset.df.copy()
+
         for col_name, col_value in list(row_filter.items()):
             if filtered_df[col_name].dtype == "object":
                 filtered_df = filtered_df[
@@ -26,7 +50,7 @@ class PredictTool(BasePredictTool):
 
         return filtered_df
 
-    def _build_from_user_input(self, feature_values: dict) -> pd.DataFrame:
+    def _get_input_from_user(self, feature_values: dict) -> pd.DataFrame:
         # Prepare background sample.
         from giskard.models.model_explanation import _get_background_example
 
@@ -42,10 +66,15 @@ class PredictTool(BasePredictTool):
         return background_sample
 
     def _prepare_input(self, feature_values: dict) -> Dataset:
-        final_input = self._filter_from_dataset(feature_values)
+        final_input = self._get_input_from_dataset(feature_values)
 
         # If no data were filtered from the dataset, build vector from the user input.
         if len(final_input) == 0:
-            final_input = self._build_from_user_input(feature_values)
+            final_input = self._get_input_from_user(feature_values)
 
         return Dataset(final_input, target=None)
+
+    def __call__(self, features_dict: dict) -> str:
+        model_input = self._prepare_input(features_dict)
+        prediction = self._model.predict(model_input).prediction
+        return ", ".join(map(str, prediction))
